backend/webserver/requirements.txt
asgiref==3.3.4
attrs==21.2.0
automat==20.2.0
autopep8==1.5.7
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
django-cors-headers==3.7.0
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django==3.2.4
djangorestframework==3.12.4
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
hypothesis==6.14.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
packaging==21.0
parsel==1.6.0
pluggy==0.13.1
priority==1.3.0
protego==0.1.16
psycopg2-binary==2.9.1
py==1.10.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydispatcher==2.0.5; platform_python_implementation == 'CPython'
pyjwt==2.1.0
pyopenssl==20.0.1
pyparsing==2.4.7
pytest-django==4.4.0
pytest==6.2.4
pytz==2021.1
queuelib==1.6.1
requests==2.25.1
scrapy==2.5.0
service-identity==21.1.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.1
toml==0.10.2
twisted==21.2.0
urllib3==1.26.5
w3lib==1.22.0
zope.interface==5.4.0
django-crontab==0.7.1